create JAVASCRIPT
How to derive a key (year) from such an array?
The idea. There is a question "What happened in the ... year".
By clicking button 1, the year is shown randomly. If the person does not know, then by clicking button 2 the answer is shown. and the picture. The answer and the picture are written in a nested array.